return values instead of logging them directly

+++ chainPattern/Numbers.java
@@ -1,22 +1,22 @@
 package chainPattern;
 
 public class Numbers {
-	private int number1;
-	private int number2;
+	private double number1;
+	private double number2;
 	private String calculationWanted;
 	
-	public Numbers(int number1, int number2, String calculationWanted) {
+	public Numbers(double number1, double number2, String calculationWanted) {
 		//super();
 		this.number1 = number1;
 		this.number2 = number2;
 		this.calculationWanted = calculationWanted;
 	}
 
-	public int getNumber1() {
+	public double getNumber1() {
 		return number1;
 	}
 
-	public int getNumber2() {
+	public double getNumber2() {
 		return number2;
 	}
